% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneids.R
\name{check_gpl}
\alias{check_gpl}
\title{Check the input gpl if is in our platform list}
\usage{
check_gpl(gpl)
}
\arguments{
\item{gpl}{}
}
\value{
returns a boolean value
}
\description{
Check the input gpl if is in our platform list
}
