% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneids.R
\name{get_anno_data}
\alias{get_anno_data}
\title{Get annotation data}
\usage{
get_anno_data(spe, biopack_status)
}
\arguments{
\item{spe, biocpack_status}{}
}
\value{
returns annotation data
}
\description{
Get annotation data
}
