% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneids.R
\name{get_geneids}
\alias{get_geneids}
\title{Mapping gene ids to probe ids}
\usage{
get_geneids(probeids, platform, plt.choose = T)
}
\arguments{
\item{probeids, platform, plt.choose}{if plt.choose is TRUE , then shows the input is gpl ,else is platform}
}
\value{
if the input dataset is in our platform list(save in gpl_list.rda) ,then the output will be a dataframe,
which includes a list of gene ids mapping to probe ids
}
\description{
\code{get_geneids} returns a list of gene ids for the input probe ids
}
\examples{
data(package = "IDmap")
probeids <- c("1000_at","1001_at","1002_f_at")
platform <- "hcg110"
geneids <- get_geneids(probeids,platform,plt.choose=F)
head(geneids)

probeids <- c("A_23_P101521","A_33_P3695548","A_33_P3266889")
platform <- "GPL10332"
geneids <- get_geneids(probeids,platform,plt.choose=T)
head(geneids)

}
