% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneids.R
\name{get_geneids_from_gpl}
\alias{get_geneids_from_gpl}
\title{Get_geneids_from_gpl}
\usage{
get_geneids_from_gpl(probeids, gpl)
}
\arguments{
\item{probeids, gpl}{}
}
\value{
if the input dataset is in our platform list(save in gpl_list.rda) ,then the output will be a dataframe,
which includes a list of gene ids mapping to probe ids
}
\description{
\code{get_geneids} returns a list of gene ids for the input probe ids
}
