% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo.R
\name{get_geo}
\alias{get_geo}
\title{Download probe annotations from GEO}
\usage{
get_geo(GPL, destDir = getwd())
}
\arguments{
\item{GPL, destDir}{}
}
\value{
{get_geo} will download probe annotations from GEO ,then the output will be a dataframe,
which includes a list of gene ids mapping to probe ids
}
\description{
\code{get_geo} returns a list of gene ids for the input GPL Accession Number
}
\examples{
get_geo("GPL13607", destdir=getwd())

}
